+++ src/main/java/nexusvault/cli/plugin/export/ExportDirCmd.java
@@ -1,0 +1,27 @@
+package nexusvault.cli.plugin.export;
+
+import nexusvault.cli.CommandArguments;
+import nexusvault.cli.CommandInfo;
+import nexusvault.cli.plugin.AbstCommand;
+
+final class ExportDirCmd extends AbstCommand {
+
+	@Override
+	public CommandInfo getCommandInfo() {
+		// TODO Auto-generated method stub
+		return null;
+	}
+
+	@Override
+	public void onCommand(CommandArguments args) {
+		// TODO Auto-generated method stub
+
+	}
+
+	@Override
+	public void onHelp(CommandArguments args) {
+		// TODO Auto-generated method stub
+
+	}
+
+}